+++ AddOns/MplusAdventureGuide/Locales/zhCN.lua
@@ -1,4 +1,0 @@
-local L = LibStub("AceLocale-3.0"):NewLocale("MplusAdventureGuide", "zhCN")
-if not L then return end
-
-L["DELVES_NEEDED_TIER8"] = "尚需完成级别8以上的地下堡 %d 场"