+++ AddOns/MplusAdventureGuide/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Mythic+ Miscellaneous
 
-## [12.0-006](https://github.com/teelolws/MplusAG/tree/12.0-006) (2026-03-25)
-[Full Changelog](https://github.com/teelolws/MplusAG/compare/12.0-005...12.0-006) [Previous Releases](https://github.com/teelolws/MplusAG/releases)
+## [12.1-001](https://github.com/teelolws/MplusAG/tree/12.1-001) (2026-08-11)
+[Full Changelog](https://github.com/teelolws/MplusAG/compare/12.0-011...12.1-001) [Previous Releases](https://github.com/teelolws/MplusAG/releases)
 
-- Add midnight season 1 portal IDs  
+- Merge branch 'main' of https://github.com/teelolws/MplusAG  
+- TOC Update for 12.1.0  
+- Merge pull request #37 from MiliChang/main  
+    Add midnight season 2 item levels  
+- Add midnight season 2 item levels  
+- Remove Premade Finder module: obsolete as of 12.1.0